% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_dark_light_theme}
\alias{auto_dark_light_theme}
\alias{set_theme_light}
\alias{set_theme_dark}
\alias{use_theme_light}
\alias{use_theme_dark}
\alias{use_theme_toggle}
\alias{use_theme_auto}
\title{Automatically or manually toggle light and dark themes}
\usage{
set_theme_light(theme = NULL)

set_theme_dark(theme = NULL)

use_theme_light(quietly = FALSE)

use_theme_dark(quietly = FALSE)

use_theme_toggle(quietly = FALSE)

use_theme_auto(dark_start = "18:00", dark_end = "6:00",
  quietly = FALSE)
}
\arguments{
\item{theme}{The name of the theme, or \code{NULL} to use current theme.}

\item{quietly}{Suppress confirmation messages}

\item{dark_start}{Start time of dark mode, in 24-hour \code{"HH:MM"} format.}

\item{dark_end}{End time of dark mode, in 24-hour \code{"HH:MM"} format.}
}
\description{
These functions help manage switching between preferred dark and light
themes. Use \code{rsthemes::set_theme_light()} and \code{rsthemes::set_theme_dark()} to
declare your preferred light/dark themes. Set the preferred theme in your
\code{~/.Rprofile} (see the section below) for ease of switching. Then use the
\code{use_theme_light()} or \code{use_theme_dark()} to switch to your preferred theme.
Switch between dark and light themes with \code{use_theme_toggle()} or
automatically set your theme to dark mode using \code{use_theme_auto()}.
}
\section{Functions}{
\itemize{
\item \code{set_theme_light}: Set default light theme

\item \code{set_theme_dark}: Set default dark theme

\item \code{use_theme_light}: Use default light theme

\item \code{use_theme_dark}: Use default dark theme

\item \code{use_theme_toggle}: Toggle between dark and light themes

\item \code{use_theme_auto}: Auto switch between dark and light themes
}}

\section{Set preferred theme in \code{.Rprofile}}{


Add the following to your \code{~/.Rprofile} (see \code{\link[usethis:edit_r_profile]{usethis::edit_r_profile()}}) to
declare your default themes:\preformatted{if (
  interactive() &&
  requireNamespace("rsthemes", quietly = TRUE) &&
  requireNamespace("later", quietly = TRUE)
) {
  # Use later to delay until RStudio is ready
  later::later(function() {
    rsthemes::set_theme_light("One Light {rsthemes}")  # light theme
    rsthemes::set_theme_dark("One Dark {rsthemes}") # dark theme

    # To automatically choose theme based on time of day
    rsthemes::use_theme_auto(dark_start = "18:00", dark_end = "6:00")
  }, delay = 1)
}
}

Note that \pkg{later} is required to add a slight delay to allow RStudio to
start up. If you'd rather not use this approach, you can simply declare the
global options that declare the default themes, but you won't be able to use
\code{\link[=use_theme_auto]{use_theme_auto()}} at startup.\preformatted{options(
  rsthemes.theme_light = "One Light {rsthemes}",
  rsthemes.theme_dark = "One Dark {rsthemes}"
)
}
}

\section{RStudio Addins}{


\pkg{rsthemes} includes four RStudio addins to help you easily switch between
light and dark modes. You can set the default dark or light theme to the
current theme. You can also toggle between light and dark mode or switch
to the automatically chosen light/dark theme based on time of day. You can
set a keyboard shortcut to \strong{Toggle Dark Mode} or
\strong{Auto Choose Dark or Light Theme} from the \emph{Modify Keyboard Shortcuts...}
window under the RSTudio \emph{Tools} menu.
}

