% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_theme}
\alias{auto_theme}
\alias{set_theme_light}
\alias{set_theme_dark}
\alias{set_theme_favorite}
\alias{use_theme_light}
\alias{use_theme_dark}
\alias{use_theme_rstudio_default}
\alias{use_theme_toggle}
\alias{use_theme_auto}
\alias{use_theme_favorite}
\title{Automatically or manually toggle light, dark, or favorite themes}
\usage{
set_theme_light(theme = NULL)

set_theme_dark(theme = NULL)

set_theme_favorite(theme = NULL, append = TRUE)

use_theme_light(quietly = FALSE)

use_theme_dark(quietly = FALSE)

use_theme_rstudio_default(quietly = FALSE)

use_theme_toggle(quietly = FALSE)

use_theme_auto(
  dark_start = NULL,
  dark_end = NULL,
  lat = NULL,
  lon = NULL,
  quietly = FALSE
)

use_theme_favorite(quietly = FALSE)
}
\arguments{
\item{theme}{The name of the theme, or \code{NULL} to use current theme.}

\item{append}{[set_theme_favorite] Should the theme be appended to the list
of favorite themes? If \code{FALSE}, then \code{theme} replaces the current list of
favorite themes.}

\item{quietly}{Suppress confirmation messages}

\item{dark_start}{Start time of dark mode, in 24-hour \code{"HH:MM"} format.}

\item{dark_end}{End time of dark mode, in 24-hour \code{"HH:MM"} format.}

\item{lat}{Latitude of your current location.}

\item{lon}{Longitude of your current location.}
}
\description{
These functions help manage switching between preferred themes. Use
\code{set_theme_light()} and \code{set_theme_dark()} to declare your preferred
light/dark themes that can be toggled or selected automatically with
\code{use_theme_toggle()} or \code{use_theme_auto()} or their corresponding RStudio
addins. Alternatively, you can create a list of favorite themes with
\code{set_theme_favorite()}. You can then cycle through this list with
\code{use_theme_favorite()} or by using the RStudio addin.
}
\details{
For best results, set your preferred themes in your \verb{~/.Rprofile} using the
instructions in the section below.
}
\section{Functions}{
\itemize{
\item \code{set_theme_light}: Set default light theme

\item \code{set_theme_dark}: Set default dark theme

\item \code{set_theme_favorite}: Set favorite themes

\item \code{use_theme_light}: Use default light theme

\item \code{use_theme_dark}: Use default dark theme

\item \code{use_theme_rstudio_default}: Use the RStudio default theme (Textmate)

\item \code{use_theme_toggle}: Toggle between dark and light themes

\item \code{use_theme_auto}: Auto switch between dark and light themes

\item \code{use_theme_favorite}: Walk through a list of favorite themes
}}

\section{Set preferred theme in \code{.Rprofile}}{


Add the following to your \verb{~/.Rprofile} (see \code{\link[usethis:edit]{usethis::edit_r_profile()}}) to
declare your default themes:\preformatted{if (interactive() && requireNamespace("rsthemes", quietly = TRUE)) \{
  # Set preferred themes if not handled elsewhere..
  rsthemes::set_theme_light("One Light \{rsthemes\}")  # light theme
  rsthemes::set_theme_dark("One Dark \{rsthemes\}") # dark theme
  rsthemes::set_theme_favorite(c(
    "GitHub \{rsthemes\}", "Solarized Dark \{rsthemes\}", "Fairyfloss \{rsthemes\}"
  ))


  # Whenever the R session restarts inside RStudio...
  setHook("rstudio.sessionInit", function(isNewSession) \{
    # Automatically choose the correct theme based on time of day
    rsthemes::use_theme_auto(dark_start = "18:00", dark_end = "6:00")

    # Alternatively use a dynamic solution based on your location
    # rsthemes::use_theme_auto(lat = 50.30, lon = 9.40)

    # Or alternatively use \{ipapi\} to determine your location automatically
    # rsthemes::use_theme_auto()
  \}, action = "append")
\}
}

Note that the fully automated approach will use
\href{https://ip-api.com/}{ip-api.com} to look up your current location using your
IP address (if the \pkg{ipapi} package is available). If you'd like to use
geolocation but without pinging an external server, you can call
\code{rsthemes:::geolocate()} once to determine your approximate latitude and
longitude, which you can provide to \code{\link[=use_theme_auto]{use_theme_auto()}}. Your location
coordinates are cached on a per-project basis and only checked once per day
(per project).

Here are two other approaches that you can use. Both set the global
options that declare the default themes, but rely on you manually calling
\code{\link[=use_theme_auto]{use_theme_auto()}} or the Auto Dark/Light addin.\preformatted{# ~/.Rprofile
rsthemes::set_theme_light("One Light \{rsthemes\}")
rsthemes::set_theme_dark("One Dark \{rsthemes\}")
rsthemes::set_theme_favorite(c(
  "GitHub \{rsthemes\}", "Solarized Dark \{rsthemes\}", "Fairyfloss \{rsthemes\}"
))
}\preformatted{# ~/.Rprofile
options(
  rsthemes.theme_light = "One Light \{rsthemes\}",
  rsthemes.theme_dark = "One Dark \{rsthemes\}"
  rsthemes.theme_favorite = c("GitHub \{rsthemes\}", "One Light \{rsthemes\}")
)
}
}

\section{RStudio Addins}{


\pkg{rsthemes} includes five RStudio addins to help you easily switch between
light and dark modes. You can set the default dark or light theme to the
current theme. You can also toggle between light and dark mode or switch
to the automatically chosen light/dark theme based on time of day. You can
set a keyboard shortcut to \strong{Toggle Dark Mode}, \strong{Next Favorite Theme}, or
\strong{Auto Choose Dark or Light Theme} from the \emph{Modify Keyboard Shortcuts...}
window under the RStudio \emph{Tools} menu.

\pkg{rsthemes} also includes \strong{Use Default RStudio Theme}, powered by
\code{use_theme_rstudio_default()}. This addin and function reverts RStudio back
to the default theme, Textmate. This can be useful when teaching or
demonstrating RStudio feature when you want RStudio to look like new.
}

