% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{rstheme}
\alias{rstheme}
\title{Create a New Theme}
\usage{
rstheme(
  theme_name,
  ui_background,
  ui_foreground,
  code_string,
  code_function,
  code_value,
  code_variable,
  code_message,
  ...,
  theme_path = NULL,
  theme_dark = FALSE,
  theme_flat = FALSE,
  theme_palette = NULL,
  theme_as_sass = FALSE,
  theme_apply = TRUE,
  ui_rstudio_background = "darken($ui_background, 2\%)",
  ui_rstudio_foreground = "$ui_foreground",
  ui_rstudio_border = "$ui_background",
  rmd_chunk_background = "darken($ui_background, 5\%)",
  ui_rstudio_tabs_inactive_background = "lighten($ui_rstudio_background, 2\%)",
  ui_rstudio_tabs_inactive_foreground = "transparentize($ui_foreground, 0.4)",
  ui_rstudio_tabs_inactive_hover_background = "$ui_rstudio_tabs_inactive_background",
  ui_rstudio_tabs_active_background = "$rmd_chunk_background",
  ui_rstudio_tabs_active_foreground = "$ui_foreground",
  ui_rstudio_toolbar_background = "$ui_rstudio_tabs_active_background",
  ui_rstudio_toolbar_foreground = "$ui_foreground",
  ui_rstudio_search = "$ui_rstudio_toolbar_background",
  ui_rstudio_scrollbar_background = "transparentize($ui_background, 0.5)",
  ui_rstudio_scrollbar_handle = "$rmd_chunk_background",
  ui_cursor = "#FF0000",
  ui_cursor_normal_mode = "$ui_cursor",
  ui_gutter_foreground = "transparentize($ui_foreground, 0.75)",
  ui_gutter_background = "$ui_background",
  ui_line_active = "transparentize($rmd_chunk_background, 0.6)",
  ui_selection = "transparentize($rmd_chunk_background, 0.6)",
  ui_line_active_selection = "opacify($ui_selection, 1)",
  ui_line_active_gutter = "$ui_background",
  ui_bracket = "transparentize($ui_cursor, 0.6)",
  ui_invisible = "transparentize($ui_foreground, 0.7)",
  ui_indent_guide = "$ui_invisible",
  ui_margin_line = "lighten($ui_selection, 5\%)",
  ui_debug_background = "lighten($ui_background, 10\%)",
  ui_fold_arrows_foreground = "$ui_cursor",
  ui_fold_arrows_background = "transparentize($ui_fold_arrows_foreground, 0.5)",
  ui_line_find = "$ui_selection",
  ui_console_selection = "$ui_selection",
  ui_completions_background = "$ui_background",
  ui_completions_foreground = "$ui_foreground",
  ui_completions_border = "$ui_bracket",
  ui_completions_selected_background = "$ui_line_active_selection",
  ui_completions_selected_foreground = "$ui_foreground",
  ui_rstudio_job_progress_bar = "opacify($ui_selection, 1)",
  code_comment = "transparentize($ui_foreground, 0.5)",
  code_namespace_font_style = "italic",
  code_reserved = "$code_string",
  code_operator = "$code_value",
  code_bracket = "$code_function",
  code_namespace = "$code_variable",
  code_identifier = "$ui_foreground",
  code_string_background = "inherit",
  code_function_background = "inherit",
  code_value_background = "inherit",
  code_variable_background = "inherit",
  code_message_background = "inherit",
  code_comment_background = "inherit",
  code_reserved_background = "inherit",
  code_operator_background = "inherit",
  code_bracket_background = "inherit",
  code_namespace_background = "inherit",
  code_identifier_background = "inherit",
  code_other = "$code_string",
  rmd_heading_weight = "600",
  rmd_heading_foreground = "$code_string",
  rmd_heading_background = "$code_string_background",
  rmd_chunk_header = "$code_function",
  rmd_chunk_header_background = "$code_function_background",
  rmd_href = "$code_reserved",
  rmd_href_background = "$code_reserved_background"
)
}
\arguments{
\item{theme_name}{Name of the theme}

\item{ui_background}{General background color in RStudio panes, e.g. Source pane.}

\item{ui_foreground}{General foreground (text) color in RStudio panes}

\item{code_string, code_function, code_value, code_variable, code_message}{Color of code tokens in R code.}

\item{...}{Raw SCSS text strings or \pkg{rstheme} partial themes:
\itemize{
\item \code{\link[=rstheme_command_palette]{rstheme_command_palette()}}
\item \code{\link[=rstheme_rainbow_parentheses]{rstheme_rainbow_parentheses()}}
\item \code{\link[=rstheme_large_tabs]{rstheme_large_tabs()}}
\item \code{\link[=rstheme_dialog_options]{rstheme_dialog_options()}}
\item \code{\link[=rstheme_terminal_colors]{rstheme_terminal_colors()}}
}}

\item{theme_path}{Path for output theme}

\item{theme_dark}{\code{TRUE} if the theme is a darke theme, \code{FALSE} otherwise
(default)}

\item{theme_flat}{Use additional classes to make the theme a "flat" theme.
Default is \code{FALSE}.}

\item{theme_palette}{A list of named colors for use in the theme palette.
Color names should not include the \code{$} prefix, e.g. \code{red}, but can be
referenced in any other color-setting variable in \code{rstheme()} as
\verb{$<color>}, e.g. \verb{$red}.}

\item{theme_as_sass}{If \code{TRUE}, the theme is returned as an SCSS file.
Otherwise, the compiled \code{.rstheme} file is returned (default)}

\item{theme_apply}{If \code{TRUE} (default), the theme will be installed and
applied immediately.}

\item{ui_rstudio_background}{Default: \verb{darken($ui_background, 2\%)}}

\item{ui_rstudio_foreground}{Default: \verb{$ui_foreground}}

\item{ui_rstudio_border}{Default: \verb{$ui_background}}

\item{rmd_chunk_background}{Default: \verb{darken($ui_background, 5\%)}}

\item{ui_rstudio_tabs_inactive_background}{Default: \verb{lighten($ui_rstudio_background, 2\%)}}

\item{ui_rstudio_tabs_inactive_foreground}{Default: \verb{transparentize($ui_foreground, 0.4)}}

\item{ui_rstudio_tabs_inactive_hover_background}{Default: \verb{$ui_rstudio_tabs_inactive_background}}

\item{ui_rstudio_tabs_active_background}{Default: \verb{$rmd_chunk_background}}

\item{ui_rstudio_tabs_active_foreground}{Default: \verb{$ui_foreground}}

\item{ui_rstudio_toolbar_background}{Default: \verb{$ui_rstudio_tabs_active_background}}

\item{ui_rstudio_toolbar_foreground}{Default: \verb{$ui_foreground}}

\item{ui_rstudio_search}{Default: \verb{$ui_rstudio_toolbar_background}}

\item{ui_rstudio_scrollbar_background}{Default: \verb{transparentize($ui_background, 0.5)}}

\item{ui_rstudio_scrollbar_handle}{Default: \verb{$rmd_chunk_background}}

\item{ui_cursor}{Default: \verb{#FF0000}}

\item{ui_cursor_normal_mode}{Default: \verb{$ui_cursor}}

\item{ui_gutter_foreground}{Default: \verb{transparentize($ui_foreground, 0.75)}}

\item{ui_gutter_background}{Default: \verb{$ui_background}}

\item{ui_line_active}{Default: \verb{transparentize($rmd_chunk_background, 0.6)}}

\item{ui_selection}{Default: \verb{transparentize($rmd_chunk_background, 0.6)}}

\item{ui_line_active_selection}{Default: \verb{opacify($ui_selection, 1)}}

\item{ui_line_active_gutter}{Default: \verb{$ui_background}}

\item{ui_bracket}{Highlight on paired bracket Default: \verb{transparentize($ui_cursor, 0.6)}}

\item{ui_invisible}{Default: \verb{transparentize($ui_foreground, 0.7)}}

\item{ui_indent_guide}{Default: \verb{$ui_invisible}}

\item{ui_margin_line}{Default: \verb{lighten($ui_selection, 5\%)}}

\item{ui_debug_background}{Default: \verb{lighten($ui_background, 10\%)}}

\item{ui_fold_arrows_foreground}{Default: \verb{$ui_cursor}}

\item{ui_fold_arrows_background}{Default: \verb{transparentize($ui_fold_arrows_foreground, 0.5)}}

\item{ui_line_find}{Default: \verb{$ui_selection}}

\item{ui_console_selection}{Default: \verb{$ui_selection}}

\item{ui_completions_background}{Default: \verb{$ui_background}}

\item{ui_completions_foreground}{Default: \verb{$ui_foreground}}

\item{ui_completions_border}{Default: \verb{$ui_bracket}}

\item{ui_completions_selected_background}{Default: \verb{$ui_line_active_selection}}

\item{ui_completions_selected_foreground}{Default: \verb{$ui_foreground}}

\item{ui_rstudio_job_progress_bar}{Default: \verb{opacify($ui_selection, 1)}}

\item{code_comment}{Default: \verb{transparentize($ui_foreground, 0.5)}}

\item{code_namespace_font_style}{Default: \code{italic}}

\item{code_reserved}{Default: \verb{$code_string}}

\item{code_operator}{Default: \verb{$code_value}}

\item{code_bracket}{Default: \verb{$code_function}}

\item{code_namespace}{Default: \verb{$code_variable}}

\item{code_identifier}{Default: \verb{$ui_foreground}}

\item{code_string_background}{Default: \code{inherit}}

\item{code_function_background}{Default: \code{inherit}}

\item{code_value_background}{Default: \code{inherit}}

\item{code_variable_background}{Default: \code{inherit}}

\item{code_message_background}{Default: \code{inherit}}

\item{code_comment_background}{Default: \code{inherit}}

\item{code_reserved_background}{Default: \code{inherit}}

\item{code_operator_background}{Default: \code{inherit}}

\item{code_bracket_background}{Default: \code{inherit}}

\item{code_namespace_background}{Default: \code{inherit}}

\item{code_identifier_background}{Default: \code{inherit}}

\item{code_other}{Default: \verb{$code_string}}

\item{rmd_heading_weight}{Default: \code{600}}

\item{rmd_heading_foreground}{Default: \verb{$code_string}}

\item{rmd_heading_background}{Default: \verb{$code_string_background}}

\item{rmd_chunk_header}{Default: \verb{$code_function}}

\item{rmd_chunk_header_background}{Default: \verb{$code_function_background}}

\item{rmd_href}{Default: \verb{$code_reserved}}

\item{rmd_href_background}{Default: \verb{$code_reserved_background}}
}
\value{
Returns the CSS or SCSS (if \code{theme_as_sass} is \code{TRUE}). If
\code{theme_path} is provided, the CSS or SCSS is written to the file and the
source is returned invisibly. If \code{theme_apply} is \code{TRUE}, the theme is also
installed and applied to the current RStudio session.
}
\description{
Writes the template files required to create a new RStudio theme. In general,
arguments with the prefix \code{theme_} refer to global settings for the built
theme. Arguments with \code{ui_rstudio_} refer to RStudio chrome such as tabs,
scroll bars, search boxes, and tool bars. Arguments starting with \code{ui_} refer
to UI elements inside the coding area, such as the cursor, selection
highlighting, etc. Arguments starting with \code{code_} refer to code syntax
colors and those starting with \code{rmd_} are specific to R Markdown syntax. The
arguments naming conventions may not be perfectly consistent (sorry!).
}
