% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{rstheme_partial}
\alias{rstheme_partial}
\alias{rstheme_command_palette}
\alias{rstheme_rainbow_parentheses}
\alias{rstheme_large_tabs}
\alias{rstheme_dialog_options}
\alias{rstheme_terminal_colors}
\title{RStudio Partial Themes}
\usage{
rstheme_command_palette(
  background = "$ui_background",
  search_background = "$ui_command_palette_background",
  border = "$ui_command_palette_background",
  item_background = "$ui_command_palette_background",
  item_color = "$ui_foreground",
 
    item_hover_background = "if($ui_rstudio_is_dark, lighten($ui_command_palette_item_background, 5\%), darken($ui_command_palette_item_background, 5\%))",
 
    item_selected_background = "if($ui_rstudio_is_dark, lighten($ui_command_palette_item_background, 10\%), darken($ui_command_palette_item_background, 10\%))"
)

rstheme_rainbow_parentheses(
  ui_paren_0 = "rgb(255, 124, 124)",
  ui_paren_1 = "rgb(255, 187, 0)",
  ui_paren_2 = "#00ff1f",
  ui_paren_3 = "#2bcf2b",
  ui_paren_4 = "lightskyblue",
  ui_paren_5 = "rgb(150, 94, 255)",
  ui_paren_6 = "violet"
)

rstheme_large_tabs()

rstheme_dialog_options(
  background = "$ui_rstudio_background",
  foreground = "$ui_rstudio_foreground",
 
    border = "if($ui_rstudio_is_dark, lighten($ui_rstudio_dialog_background, 5\%), darken($ui_rstudio_dialog_background, 5\%))",
  selected_background = "$ui_rstudio_dialog_border",
  selected_foreground = "$ui_rstudio_dialog_foreground",
  heading_foreground = "$code_function",
  help_foreground = "$code_value",
  button_foreground = "$ui_rstudio_dialog_foreground",
  button_background = "$ui_rstudio_dialog_background",
  button_border = "$ui_rstudio_dialog_border",
  button_hover_foreground = "lighten($ui_rstudio_dialog_button_foreground, 10\%)",
  button_hover_background = "lighten($ui_rstudio_dialog_button_background, 10\%)",
  button_hover_border = "$ui_rstudio_dialog_border",
  input_foreground = "$ui_rstudio_dialog_foreground",
  input_background = "$ui_rstudio_dialog_background",
  input_border = "$ui_rstudio_dialog_border",
  checkbox_background = "lighten($ui_rstudio_dialog_background, 5\%)",
  checkbox_foreground = "$ui_rstudio_dialog_foreground",
  select_background = "lighten($ui_rstudio_dialog_background, 5\%)",
  select_foreground = "$ui_rstudio_dialog_foreground"
)

rstheme_terminal_colors(
  theme_dark = TRUE,
  red = NULL,
  green = NULL,
  yellow = NULL,
  blue = NULL,
  magenta = NULL,
  cyan = NULL,
  black = if (theme_dark) "darken($ui_background, 5\%)" else
    "darken($ui_foreground, 5\%)",
  white = if (theme_dark) "darken($ui_foreground, 5\%)" else
    "darken($ui_background, 5\%)",
  red_bright = red,
  green_bright = green,
  yellow_bright = yellow,
  blue_bright = blue,
  magenta_bright = magenta,
  cyan_bright = cyan,
  black_bright = "mix($terminal_color_black, $terminal_color_white, 40\%)",
  white_bright = "lighten($terminal_color_white, 10\%)"
)
}
\arguments{
\item{search_background, item_background, item_color, item_hover_background, item_selected_background}{Additional color options for the command palette.}

\item{border}{Border color}

\item{ui_paren_0, ui_paren_1, ui_paren_2, ui_paren_3, ui_paren_4, ui_paren_5, ui_paren_6}{Parentheses colors 0 through 6.}

\item{foreground, background}{Foreground (text) or background color}

\item{selected_background, selected_foreground, heading_foreground, help_foreground, button_foreground, button_background, button_border, button_hover_foreground, button_hover_background, button_hover_border, input_foreground, input_background, input_border, checkbox_background, checkbox_foreground, select_background, select_foreground}{Additional colors for RStudio dialog boxes, setting SCSS variables prefixed
with \verb{$ui_rstudio_dialog_<argument>}.}

\item{theme_dark}{Is the theme dark (otherwise light). Used to set the
default terminal colors for black and white based on the theme's foreground
and background colors.}

\item{black, black_bright, red, red_bright, green, green_bright, yellow, yellow_bright, blue, blue_bright, magenta, magenta_bright, cyan, cyan_bright, white, white_bright}{Terminal colors, 8 normal and 8 bright.}
}
\description{
Partial themes for RStudio to be included as desired in \code{\link[=rstheme]{rstheme()}}.
}
\section{Functions}{
\itemize{
\item \code{rstheme_command_palette}: Theme the RStudio Command Palette

\item \code{rstheme_rainbow_parentheses}: Theme the rainbow parentheses colors

\item \code{rstheme_large_tabs}: Enlarges the tabs

\item \code{rstheme_dialog_options}: Themes dialog windows, e.g. \emph{Global Options}

\item \code{rstheme_terminal_colors}: Theme the terminal colors
}}

