% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsthemes-package.R
\docType{package}
\name{rsthemes-package}
\alias{rsthemes}
\alias{rsthemes-package}
\title{rsthemes: Full Themes for RStudio v1.2+}
\description{
This package provides a set of complete RStudio themes that restyle the
IDE and the editor syntax. The themes are built from
\href{https://sass-lang.org}{sass} templates into \code{.rstheme} files that,
ultimately, contain the CSS files that provide the themes. The sass templates
provide an easy-to-extend framework for customizing the look of RStudio. If
you would like to create your own theme, \code{\link[=rstheme]{rstheme()}} is designed to help you
get started quickly.
}
\section{Getting Started}{

\itemize{
\item Would you like to install, try out, or remove RStudio themes from
\pkg{rsthemes}? Read more in \link{rsthemes-actions}.
\item Would you like to automatically choose a dark or light theme? Or would
you like to quickly toggle through a set of themes you enjoy? Read more
in \link{auto_theme}.
\item Would you like to create a new theme for RStudio? Read more in
\code{\link[=rstheme]{rstheme()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/rsthemes}
  \item Report bugs at \url{https://github.com/gadenbuie/rsthemes/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
