% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsthemes.R
\name{rsthemes}
\alias{rsthemes}
\alias{install_rsthemes}
\alias{remove_rsthemes}
\alias{list_rsthemes}
\alias{try_rsthemes}
\alias{rsthemes_styles}
\title{Full RStudio Themes}
\usage{
install_rsthemes(style = "all", include_base16 = FALSE, destdir = NULL)

remove_rsthemes(style = rsthemes_styles(), include_base16 = TRUE)

list_rsthemes(style = "all", include_base16 = TRUE, list_installed = TRUE)

try_rsthemes(style = "all", include_base16 = TRUE, delay = 0)

rsthemes_styles(..., validate = NULL)
}
\arguments{
\item{style}{Limit to a subgroup of themes, chosen from the options returned
by \code{\link[=rsthemes_styles]{rsthemes_styles()}}.}

\item{include_base16}{Should the \code{base16} themes be included?}

\item{destdir}{The destination directory for the \code{.rstheme} files. By default
uses \code{\link[rstudioapi:addTheme]{rstudioapi::addTheme()}} to install themes, but this argument lets
users install themes to non-standard directories, or in case the location
of the RStudio theme directory has changed.}

\item{list_installed}{Should the installed \pkg{rsthemes} themes be listed
(default). If \code{FALSE}, the available themes in the \pkg{rsthemes} package
are listed instead.}

\item{delay}{Number of seconds to wait between themes. Set to 0 to be
prompted to continue after each theme.}

\item{...}{Ignored.}

\item{validate}{If provided to \code{rsthemes_styles()}, checks that the
specified style is valid.}
}
\description{
This package provides a set of complete RStudio themes that restyle the
IDE and the editor syntax. The themes are built from
\href{https://sass-lang.org}{sass} templates into \code{.rstheme} files that,
ultimately, contain the CSS files that provide the themes. The sass templates
provide an easy-to-extend framework for customizing the look of RStudio. See
\code{\link[=rstheme]{rstheme()}} for more information about creating your own theme. I've done
my best to avoid hacky CSS rules wherever possible, and the themes are tested
against the current production release of RStudio (version 1.2.1335 through
1.4+ as of April 01, 2021). Please report any issues you encounter here:
\href{https://github.com/gadenbuie/rsthemes/issues}{github.com/gadenbuie/rsthemes/issues}.
}
\section{Functions}{
\itemize{
\item \code{install_rsthemes}: Install RStudio themes

\item \code{remove_rsthemes}: Remove rsthemes from RStudio

\item \code{list_rsthemes}: List installed themes (default) or available themes

\item \code{try_rsthemes}: Try each rsthemes RStudio theme

\item \code{rsthemes_styles}: List style options
}}

\section{Palettes}{
 \pkg{rsthemes} includes RStudio themes based on the
following colors palettes.
\itemize{
\item \emph{\href{https://github.com/chriskempson/base16}{base16}} (Various Authors)
\item \emph{\href{https://github.com/sailorhg/fairyfloss}{Fairyfloss}} (\href{https://github.com/sailorhg}{Amy Wibowo (sailorhg)})
\item \emph{\href{https://github.com/biletskyy/flatwhite-syntax}{Flat White}} (\href{https://github.com/biletskyy}{Dmitry Biletskyy})
\item \emph{\href{https://github.com/arcticicestudio/nord}{Nord}} (\href{https://www.svengreb.de/}{Sven Greb})
\item \emph{\href{https://github.com/marcoms/oceanic-plus}{Oceanic Plus}} (\href{https://github.com/marcoms}{Marco Scannadinari})
\item \emph{\href{https://github.com/atom/atom/tree/master/packages/one-dark-syntax}{Atom One Dark}}
\item \emph{\href{https://github.com/atom/atom/tree/master/packages/one-light-syntax}{Atom One Light}}
\item \emph{\href{https://ethanschoonover.com/solarized}{Solarized}} (Ethan Schoonover)
\item \emph{\href{https://horizontheme.netlify.app/}{Horizon Dark}} (Jonathan Olaleye)
\item \emph{\href{https://github.com/ericwbailey/a11y-syntax-highlighting}{a11y-syntax-highlighting}} (\href{https://ericwbailey.design/}{Eric Bailey})
\item \emph{\href{https://github.com/sdras/night-owl-vscode-theme}{Night Owl}} (\href{https://sarah.dev/}{Sarah Drasner})
\item \emph{\href{https://material-theme.site/}{Material Themes}}
}
}

