% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsthemes.R
\name{rsthemes}
\alias{rsthemes}
\alias{install_rsthemes}
\alias{remove_rsthemes}
\alias{list_rsthemes}
\alias{try_rsthemes}
\alias{rsthemes_styles}
\title{Full RStudio Themes}
\usage{
install_rsthemes(style = "all", include_base16 = TRUE)

remove_rsthemes(style = rsthemes_styles(), include_base16 = TRUE)

list_rsthemes(style = "all", include_base16 = TRUE,
  list_installed = TRUE)

try_rsthemes(style = "all", include_base16 = TRUE, delay = 0)

rsthemes_styles(..., validate = NULL)
}
\arguments{
\item{style}{Limit to a subgroup of themes, chosen from the options returned
by \code{\link[=rsthemes_styles]{rsthemes_styles()}}.}

\item{include_base16}{Should the \code{base16} themes be included?}

\item{list_installed}{Should the installed \pkg{rsthemes} themes be listed
(default). If \code{FALSE}, the available themes in the \pkg{rsthemes} package
are listed instead.}

\item{delay}{Number of seconds to wait bewtween themes. Set to 0 to be
prompted to continue after each theme.}
}
\description{
This package provides a set of complete RStudio themes that restyle the
IDE and the editor syntax. The themes are built from
\href{https://sass-lang.org}{sass} templates into \code{.rstheme} files that,
ultimately, contain the CSS files that provide the themes. The sass templates
provide an easy-to-extend framework for customizing the look of RStudio. See
\code{\link[=new_rstheme]{new_rstheme()}} for more information about creating your own theme. I've done
my best to avoid hacky CSS rules wherever possible, and the themes are tested
against the current production release of RStudio (verision 1.2.1335 as of
June 11, 2019). Please report any issues you encounter here:
\href{https://github.com/gadenbuie/rsthemes/issues}{github.com/gadenbuie/rsthemes/issues}.
}
\section{Functions}{
\itemize{
\item \code{install_rsthemes}: Install RStudio themes

\item \code{remove_rsthemes}: Remove rsthemes from RStudio

\item \code{list_rsthemes}: List installed themes (default) or available themes

\item \code{try_rsthemes}: Try each rsthemes RStudio theme

\item \code{rsthemes_styles}: List style options
}}

