% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsthemes.R
\name{rsthemes}
\alias{rsthemes}
\alias{install_rsthemes}
\alias{remove_rsthemes}
\alias{list_rsthemes}
\alias{try_rsthemes}
\alias{rsthemes_styles}
\title{Full RStudio Themes}
\usage{
install_rsthemes(style = "all", include_base16 = FALSE, destdir = NULL)

remove_rsthemes(style = rsthemes_styles(), include_base16 = TRUE)

list_rsthemes(style = "all", include_base16 = TRUE, list_installed = TRUE)

try_rsthemes(style = "all", include_base16 = TRUE, delay = 0)

rsthemes_styles(..., validate = NULL)
}
\arguments{
\item{style}{Limit to a subgroup of themes, chosen from the options returned
by \code{\link[=rsthemes_styles]{rsthemes_styles()}}.}

\item{include_base16}{Should the \code{base16} themes be included?}

\item{destdir}{The destination directory for the \code{.rstheme} files. By default
uses \code{\link[rstudioapi:addTheme]{rstudioapi::addTheme()}} to install themes, but this argument lets
users install themes to non-standard directories, or in case the location
of the RStudio theme directory has changed.}

\item{list_installed}{Should the installed \pkg{rsthemes} themes be listed
(default). If \code{FALSE}, the available themes in the \pkg{rsthemes} package
are listed instead.}

\item{delay}{Number of seconds to wait between themes. Set to 0 to be
prompted to continue after each theme.}

\item{...}{Ignored.}

\item{validate}{If provided to \code{rsthemes_styles()}, checks that the
specified style is valid.}
}
\description{
This package provides a set of complete RStudio themes that restyle the
IDE and the editor syntax. The themes are built from
\href{https://sass-lang.org}{sass} templates into \code{.rstheme} files that,
ultimately, contain the CSS files that provide the themes. The sass templates
provide an easy-to-extend framework for customizing the look of RStudio. If
you would like to create your own theme, \code{\link[=rstheme]{rstheme()}} is designed to help you
get started quickly.
}
\section{Functions}{
\itemize{
\item \code{install_rsthemes}: Install RStudio themes

\item \code{remove_rsthemes}: Remove rsthemes from RStudio

\item \code{list_rsthemes}: List installed themes (default) or available themes

\item \code{try_rsthemes}: Try each rsthemes RStudio theme

\item \code{rsthemes_styles}: List style options
}}

\section{Palettes}{
 \pkg{rsthemes} includes RStudio themes based on the
following colors palettes.
\itemize{
\item \href{https://github.com/chriskempson/base16}{base16} (Various Authors)
\item \href{https://github.com/sailorhg/fairyfloss}{Fairyfloss} (\href{https://github.com/sailorhg}{Amy Wibowo (sailorhg)})
\item \href{https://github.com/biletskyy/flatwhite-syntax}{Flat White} (\href{https://github.com/biletskyy}{Dmitry Biletskyy})
\item \href{https://github.com/arcticicestudio/nord}{Nord} (\href{https://www.svengreb.de/}{Sven Greb})
\item \href{https://github.com/marcoms/oceanic-plus}{Oceanic Plus} (\href{https://github.com/marcoms}{Marco Scannadinari})
\item \href{https://github.com/atom/atom/tree/master/packages/one-dark-syntax}{Atom One Dark}
\item \href{https://github.com/atom/atom/tree/master/packages/one-light-syntax}{Atom One Light}
\item \href{https://ethanschoonover.com/solarized}{Solarized} (Ethan Schoonover)
\item \href{https://horizontheme.netlify.app/}{Horizon Dark} (Jonathan Olaleye)
\item \href{https://github.com/ericwbailey/a11y-syntax-highlighting}{a11y-syntax-highlighting} (\href{https://ericwbailey.design/}{Eric Bailey})
\item \href{https://github.com/sdras/night-owl-vscode-theme}{Night Owl} (\href{https://sarah.dev/}{Sarah Drasner})
\itemize{
\item with huge thanks to original \href{https://github.com/batpigandme/night-owlish}{Night Owlish}
implementation in RStudio by \href{https://maraaverick.rbind.io/}{Mara Averick}
}
\item Yule RStudio
\itemize{
\item Based on the \href{https://tmtheme-editor.herokuapp.com/#!/editor/theme/Yule}{Yule tmTheme}
\item Ported from \href{https://github.com/gadenbuie/yule-rstudio}{gadenbuie/yule-rstudio}
\item Featuring a background image by \href{https://unsplash.com/@joannakosinska}{Joanna Kosinska}
}
\item \href{https://material-theme.site/}{Material Theme}
\itemize{
\item Contributed to rsthemes by \href{https://www.zacdelusignan.com/}{Zac de Lusignan}
}
\item \href{https://wvsc.dev/}{Serendipity} (\href{https://www.wickedtemplates.com/}{wickedtemplates})
}
}

